requirements.txt
gradio==4.41.0
tensorflow==2.13.0
numpy
pillow
